requirements.txt
agents
python-dotenv
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
openai>=1.40.0
pymilvus>=2.4.5
numpy>=1.24
